Refactor each member of the schema into multiple files
