+++ server/src/schema/center/fields.js
@@ -1,0 +1,3 @@
+export const createdBy = (obj, args, { dataLoaders: { Users } }, info) => {
+  return Users.load(obj.createdBy)
+}
