+++ server/src/schema/center/mutations.js
@@ -1,0 +1,3 @@
+export const addCenter = (obj, args, { collections: { Centers } }, info) => {
+  return true
+}
