+++ server/src/schema/center/queries.js
@@ -1,0 +1,3 @@
+export const centers = (obj, args, { collections: { Centers } }, info) => {
+  return Centers.find().toArray()
+}
