+++ server/src/schema/user/queries.js
@@ -1,0 +1,5 @@
+export const whoami = (obj, args, context, info) => {
+  const { id, email } = context.user || {}
+
+  return id ? `${id}: ${email}` : `anonymous`
+}
